Rename datasource file
