+++ src/components/query/QueryEditor.tsx
@@ -1,7 +1,7 @@
 import defaults from 'lodash/defaults';
 import React from 'react';
 import { QueryEditorProps, SelectableValue } from '@grafana/data';
-import { DataSource } from 'DataSource';
+import { DataSource } from 'SitewiseDataSource';
 import { SitewiseQuery, SitewiseOptions, QueryType, ListAssetsQuery, ListTimeSeriesQuery } from 'types';
 import { Icon, LinkButton, Select } from '@grafana/ui';
 import { QueryTypeInfo, siteWiseQueryTypes, changeQueryType } from 'queryInfo';
@@ -60,8 +60,8 @@ export function QueryEditor(props: Props) {
         return null; // nothing required
       case QueryType.ListAssets:
         return <ListAssetsQueryEditor {...props} query={query as ListAssetsQuery} />;
-        case QueryType.ListTimeSeries:
-        return <ListTimeSeriesQueryEditorFunction {...props} query={query as ListTimeSeriesQuery} />
+      case QueryType.ListTimeSeries:
+        return <ListTimeSeriesQueryEditorFunction {...props} query={query as ListTimeSeriesQuery} />;
       case QueryType.ListAssociatedAssets:
       case QueryType.PropertyValue:
       case QueryType.PropertyInterpolated:
