+++ src/components/query/types.ts
@@ -1,4 +1,4 @@
-import { DataSource } from 'DataSource';
+import { DataSource } from 'SitewiseDataSource';
 import { SitewiseQuery } from 'types';
 
 export interface SitewiseQueryEditorProps<TQuery extends SitewiseQuery = SitewiseQuery> {
