+++ src/variables.ts
@@ -2,7 +2,7 @@ import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { assign } from 'lodash';
 import { QueryType, SitewiseQuery } from './types';
-import { DataSource } from './DataSource';
+import { DataSource } from './SitewiseDataSource';
 import { DataQueryRequest, DataQueryResponse, CustomVariableSupport, DataFrameView } from '@grafana/data';
 import { QueryEditor } from './components/query/QueryEditor';
 import { AssetModelSummary } from 'queryResponseTypes';
@@ -25,28 +25,28 @@ export class SitewiseVariableSupport extends CustomVariableSupport<DataSource, S
       case QueryType.ListAssociatedAssets:
         return this.parseOptions(response);
       default:
-        return response
+        return response;
     }
   }
 
-  parseOptions(response: Observable<DataQueryResponse> ): Observable<DataQueryResponse> {
+  parseOptions(response: Observable<DataQueryResponse>): Observable<DataQueryResponse> {
     return response.pipe(
       map((res) => {
-        let data = []
+        let data = [];
         if (res.data.length) {
-          data = res.data[0]
+          data = res.data[0];
         }
-        return {data: new DataFrameView<AssetModelSummary>(data)};
+        return { data: new DataFrameView<AssetModelSummary>(data) };
       }),
       map((res) => {
-        const newData = res.data.map((m)=>{
+        const newData = res.data.map((m) => {
           return {
-          value: m.id,
-          text: m.name,
-        }})
-        return {data:newData}
+            value: m.id,
+            text: m.name,
+          };
+        });
+        return { data: newData };
       })
-    )
+    );
   }
-
 }
